Fix hunger status bar minus icon disappearing after a scene load
The icon used to signify that calories are being transferred out
of body fat and into the hunger bar used to disappear after a
scene load. This commit fixes this in the silliest way possible,
by just creating the minus sign by taking another rectangular icon
and scaling it down to a minus-shaped box.